pyproject.toml
[tool.poetry]
name = "task-management-system"
version = "0.1.0"
description = ""
authors = ["Angelina Pan <angln.pnn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pre-commit = "^4.0.1"
ruff = "^0.9.1"
redis = "^5.2.1"
celery = "^5.4.0"
alembic = "^1.14.0"
pydantic-settings = "^2.7.1"
aiologger = "^0.7.0"
pydentic = "^0.0.1.dev3"
greenlet = "^3.1.1"
asyncpg = "^0.30.0"
sqlalchemy-utils = "^0.41.2"
sqlalchemy-views = "^0.3.2"
fastapi-jwt = "^0.3.0"
psycopg2-binary = "^2.9.10"
email-validator = "^2.2.0"
jinja2 = "^3.1.5"
sqladmin = "^0.20.1"
bcrypt = "4.0.1"
httpx = "^0.28.1"
requests-oauthlib = "^2.0.0"
authlib = "^1.4.1"
flower = "^2.0.1"
fastapi-mail = "^1.4.2"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-mock = "^3.14.0"
alembic-postgresql-enum = "^1.6.0"
fastapi-filter = "^2.0.1"
allure-pytest = "^2.13.5"
passlib = "^1.7.4"
python-slugify = "^8.0.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
